requirements.txt
streamlit
pandas
matplotlib
seaborn
langchain
langchain-community
langchain-google-genai
langchain-experimental
tabulate
fpdf2